tools/nac-ise/pyproject.toml
[tool.poetry]
name = "nac-ise"
version = "0.1.0"
description = ""
authors = ["Kuba Mazurkiewicz <kmazurki@cisco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
openpyxl = "^3.1.2"
pyyaml = "^6.0.1"
argparse = "^1.4.0"

[tool.poetry.scripts]
nac-ise = "nac_ise.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
